esdf改为有符号,同时增加并行化

+++ include/rose_map/common.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include "yaml-cpp/yaml.h"
 #include <Eigen/Dense>
-using Clock = float;
+using Clock = double;
